merge

+++ notebooks/merge.py
@@ -14,7 +14,7 @@ DEEPGLOBE_ROOT = '../data/deepglobe'
 MERGED_ROOT = '../data/merge'
 
 
-def map_class_to_color(image_path, output_path):
+def map_color_to_class(image_path, output_path):
     # Open the image
     img = Image.open(image_path)
 
@@ -145,17 +145,71 @@ def split_images(dir, OUTPUT_DIR="../data/merge"):
                     # if os.path.exists(out_color_mask_path):
                     #     print(f"The mask-file {out_color_mask_path} already exists.")
                     #     continue
-                    cv2.imwrite(out_color_mask_path, mask_tile)
+                    if os.path.exists(out_mask_path):
+                        print(f"The mask-file {out_mask_path} already exists.")
+                        continue
+                    if landcoverai:
+                        cv2.imwrite(out_mask_path, mask_tile)
+
+                        map_class_to_color(out_mask_path, out_color_mask_path)
                     if os.path.exists(out_mask_path):
                         print(f"The mask-file {out_mask_path} already exists.")
                         continue
                     if deepglobe:
+                        cv2.imwrite(out_color_mask_path, mask_tile)
                         # map class to color
-                        map_class_to_color(out_color_mask_path, out_mask_path)
+                        map_color_to_class(out_color_mask_path, out_mask_path)
                 k += 1
 
         print("Processed {} {}/{}".format(img_filename, i + 1, len(img_paths)))
     check_pixel_size(OUTPUT_DIR)
+
+
+def map_class_to_color(image_path, output_path):
+    # Open the image
+    img = Image.open(image_path)
+
+    # Get the width and height of the image
+    width, height = img.size
+
+    turquoise = (0, 255, 255)
+    green = (0, 255, 0)
+    blue = (0, 0, 255)
+    black = (0, 0, 0)
+    building = 1
+    woodland = 2
+    water = 3
+    road = 4
+    background = 0
+
+    # Iterate over each pixel in the image
+    for x in range(width):
+        for y in range(height):
+            # Get the RGB values of the pixel
+            r, g, b = img.getpixel((x, y))
+
+            # Check if the pixel values fall within the specified range (1 to 5)
+            if 0 <= r <= 5 and 0 <= g <= 5 and 0 <= b <= 5:
+                if r == g and g == b:
+                    if r == building:
+                        img.putpixel((x, y), turquoise)
+                    elif r == woodland:
+                        img.putpixel((x, y), green)
+                    elif r == water:
+                        img.putpixel((x, y), blue)
+                    elif r == road:
+                        img.putpixel((x, y), black)
+                    elif r == background:
+                        img.putpixel((x, y), black)
+                    else:
+                        print(f"not a valid class. Pixel: ({r},{g},{b})")
+                        return
+                else:
+                    print(f"not a number between 1 and 5. Pixel: ({r},{g},{b})")
+                    return
+
+    # Save the modified image
+    img.save(output_path)
 
 
 def check_pixel_size(folder_path):
